requirements.txt
# Core app + UI
streamlit==1.36.0

# LangChain core + integrations
langchain==0.2.13
langchain-community==0.2.12

# Hugging Face SDKs
transformers==4.34.0
sentence-transformers==2.2.2
torch==2.2.0

# Vector store
chromadb==0.5.5

# PDF parsing
pypdf==4.3.1
PyPDF2==3.0.1

# Tokenization
tiktoken==0.7.0

# Env management
python-dotenv==1.0.1

# Utilities
uvicorn==0.30.1
typing_extensions==4.12.2
